* main.h, keygen.c (parse_expire_string, ask_expire_interval), sign.c (sign_file, clearsign_file, sign_symencrypt_file), g10.c (main), keyedit.c (sign_uids): Use seconds rather than days internally to calculate expiration.  We no longer need the day-based code as we don't generate v3 keys.

+++ g10/g10.c
@@ -2234,7 +2234,7 @@ main( int argc, char **argv )
 	  case oDefSigExpire:
 	    if(*pargs.r.ret_str!='\0')
 	      {
-		if(parse_expire_string(pargs.r.ret_str)==-1)
+		if(parse_expire_string(pargs.r.ret_str)==(u32)-1)
 		  log_error(_("`%s' is not a valid signature expiration\n"),
 			    pargs.r.ret_str);
 		else
@@ -2246,7 +2246,7 @@ main( int argc, char **argv )
 	  case oDefCertExpire:
 	    if(*pargs.r.ret_str!='\0')
 	      {
-		if(parse_expire_string(pargs.r.ret_str)==-1)
+		if(parse_expire_string(pargs.r.ret_str)==(u32)-1)
 		  log_error(_("`%s' is not a valid signature expiration\n"),
 			    pargs.r.ret_str);
 		else
